add unwrap_cast module to simplify expressions
